First bug fix is that the "unshadow" token conversion mistakenly transforms the oldest shadow token back into a shadow token.
Second bug fix is that callee nodes from an allocation site don't use the parameter rewrite rules (H) and similarly edges in the callee that are added to the caller don't use the J or K rewrite rules either.  Those reachability sets should use their callee alpha() or beta() as the input "rules" to the rewrite method.