Under Original/Normal_Java/ one would find the original Labyrinth project ported to Transactional Memory by Jihoon and reported to normal Java by Stephen Yang
Under mlp is the Labyrinth project modified with rblocks. Under mlp/Normal_Java is a version that can be compiled and run on the normal Java platform. Under rBlocked is the rBlocked version that SHOULD run under OoOJava compiler. Both of which are ported by Stephen Yang (from Jihoon's original port)

NOTE that at this point in time, the rBlocked version causes a NullPointerException in the OoOJava compiler but will compile and produce correct results under normal Java.