Adding Labyrinth3D files
in mlp, Normal_Java is a form of Labyrinth3D that can compile under normal Java.
in mlp, rBlocked is the rBlocked version that will compile with our compiler
NOTE: there seems to be a bug with make java in that the program skips right over rBlocks to print out statistics on the last part of the program EVEN THOUGH there are data dependencies. Yonghun suspects it's a bug in the compiler and will investigate over the weekend

in Original, Normal_Java is the original Labyrinth3D code with TM code removed and slightly modded to work with Normal Java.