The compiler is BROKEN, but it's NOT MY FAULT. I think Brian forgot to check in something. Anyway, now my code can do single imports and on demand multi-imports.
KNOWN PROBLEM: The package declarations can go only 1 level deep.. this means that imports would only work 1 level deep as well (but the logic for it is all correct). I think the problem lies in the .cup file... When I get the name of a package, I only get the last thing in the list of something.something.something.last.class when I ask for the package. I'll check it out later.