it passes the definitely written analysis.
the current implementation of the synthesis filter assumes to have the previous internal array values of the previous granule when it combines the subbands to the original PCM signal. So it calls for introducing the temporal data structure keeping the previous ones and making it fall down once per a loop. seems to sacrifice memory spaces for desireable properties but still looking for a better idea to handle it.